Updated BlockNavigation component
